+++ neurom/features/morphology.py
@@ -684,3 +684,41 @@ def shape_factor(morph, neurite_type=NeuriteType.all, projection_plane="xy"):
     """
     projected_points = _unique_projected_points(morph, projection_plane, neurite_type)
     return np.nan if len(projected_points) == 0 else morphmath.shape_factor(projected_points)
+
+
+@feature(shape=())
+def length_fraction_above_soma(morph, neurite_type=NeuriteType.all, axis="Y"):
+    """Returns the length fraction of the segments that have their midpoints higher than the soma.
+
+    Args:
+        morph: Morphology object.
+        neurite_type: The neurite type to use. By default all neurite types are used.
+        axis: The axis along which the computation is performed. One of ('X', 'Y', 'Z').
+
+    Returns:
+        The fraction of neurite length that lies on the right of the soma along the given axis.
+    """
+    axis = axis.upper()
+
+    if axis not in ("X", "Y", "Z"):
+        raise NeuroMError(f"Unknown axis {axis}. Please choose 'X', 'Y', or 'Z'.")
+
+    col = getattr(COLS, axis)
+    segments = list(iter_segments(morph, neurite_filter=is_type(neurite_type)))
+
+    if not segments:
+        return np.nan
+
+    # (Segment 1, Segment 2) x (X, Y, Z, R) X N
+    segments = np.dstack(segments)
+
+    # shape N x 3
+    seg_begs = segments[0, COLS.XYZ, :].T
+    seg_ends = segments[1, COLS.XYZ, :].T
+
+    lengths = np.linalg.norm(seg_begs - seg_ends, axis=1)
+
+    midpoints = 0.5 * (seg_begs + seg_ends)
+    selection = midpoints[:, col] > morph.soma.center[col]
+
+    return lengths[selection].sum() / lengths.sum()